#164 Refactoring
