+++ src/components/modal.html
@@ -1,5 +1,5 @@
 <template>
-  <require from="../components/markdown"></require>
+  <require from="../resources/custom-elements/markdown"></require>
    <div tabindex="-1" role="dialog">
       <div class="modal-dialog modal-lg">
          <div class="modal-content">
