+++ src/favorites/favorites.js
@@ -1,25 +1,25 @@
-import { inject, bindable } from 'aurelia-framework';
+// TODO Non-functional placeholder replica of most popular
+
+import { inject } from 'aurelia-framework';
 import { Router } from 'aurelia-router';
 import { DataContext } from '../services/datacontext';
 import { Filters } from '../components/filters';
 
 @inject(DataContext, Router, Filters)
-export class Explore {
+export class Favorites {
+
+  projectTitle = 'Favorite Projects';
 
   constructor(dataContext, router, filters) {
     this.dataContext = dataContext;
     this.router = router;
     this.filters = filters;
 
-    this.projectTitle = 'Explore';
-
     this.projects = [];
-    this.selectedOrganizations = [];
-    this.selectedLanguages = [];
 
     this.sortDirection = 'descending';
 
-    this.selectedSort = 'stars';
+    this.selectedSort = 'rank';
     this.sortOptions = [
       { value: 'rank', name: 'Rank' },
       { value: 'stars', name: 'Stars' },
@@ -30,25 +30,17 @@ export class Explore {
     ];
   }
 
-  gotoProject(project) {
-    this.router.navigateToRoute('edit', { id: project.id });
-  }
-
-  new() {
-    this.router.navigateToRoute('create');
-  }
-
   getData() {
-    return this.dataContext.getAll()
-      .then(projects => {
-        this.projects = JSON.parse(JSON.stringify(projects));
-        this.filters.selectedOrganizations = this.filters.getUniqueValues(this.projects, 'organization');
-        this.filters.selectedLanguages = this.filters.getUniqueValues(this.projects, 'language');
-        return this.projects;
-      });
+    return this.dataContext.getPopular().then(results => {
+      this.projects = results;
+      this.filters.selectedOrganizations = this.filters.getUniqueValues(this.projects, 'organization');
+      this.filters.selectedLanguages = this.filters.getUniqueValues(this.projects, 'language');
+      return this.projects;
+    });
   }
 
   activate() {
     this.getData();
   }
+
 }
