+++ src/nav-bar.html
@@ -2,7 +2,7 @@
 <nav class="navbar navbar-default">
   <div class="container-fluid">
       <div class="navbar-header">
-          <a class="navbar-brand" route-href="route:project">
+          <a class="navbar-brand" route-href="route:popular">
             <img src="../../../../img/stage-logo-color.svg" alt="Stage" id="logo">
             <!-- <img src="../../../../img/stage-logo-white.svg" alt="Stage" id="logo"> -->
           </a>
