+++ src/popular/popular.html
@@ -1,8 +1,8 @@
 <template>
-  <require from="../utility/sort"></require>
-  <require from="../utility/pick"></require>
-  <require from="../utility/filter"></require>
-  <require from="../utility/unique"></require>
+  <require from="../resources/value-converters/sort"></require>
+  <require from="../resources/value-converters/pick"></require>
+  <require from="../resources/value-converters/filter"></require>
+  <require from="../resources/value-converters/unique"></require>
 
   <div class="container-fluid title-wrapper">
     <h2 class="page-title"><span>${projectTitle}</span></h2>
