+++ src/project-details/project-details.html
@@ -1,6 +1,6 @@
 <template>
     <require from="../search/search-bar"></require>
-    <require from="../utility/pick"></require>
+    <require from="../resources/value-converters/pick"></require>
 
     <div class="container-fluid title-wrapper">
         <h2 class="page-title"><span>${repo.project_name}</span></h2>
